requirements.txt
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
fonttools==4.37.4
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.1
mypy==0.982
mypy-extensions==0.4.3
neo4j==5.1.0
numpy==1.23.4
packaging==21.3
pandas==1.5.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.2
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
